+++ src/NEventStore/Serialization/SerializationExtensions.cs
@@ -13,7 +13,7 @@ namespace NEventStore.Serialization
         /// <param name="serializer">The serializer to use.</param>
         /// <param name="value">The object graph to be serialized.</param>
         /// <returns>A serialized representation of the object graph provided.</returns>
-        public static byte[] Serialize<T>(this ISerialize serializer, T value)
+        public static byte[] Serialize<T>(this ISerialize serializer, T value) where T : notnull
         {
             using var stream = new MemoryStream();
             serializer.Serialize(stream, value);
@@ -27,7 +27,7 @@ namespace NEventStore.Serialization
         /// <param name="serializer">The serializer to use.</param>
         /// <param name="serialized">The serialized array of bytes.</param>
         /// <returns>The reconstituted object, if any.</returns>
-        public static T Deserialize<T>(this ISerialize serializer, byte[] serialized)
+        public static T? Deserialize<T>(this ISerialize serializer, byte[] serialized)
         {
             // add null or empty check
             if (serialized == null || serialized.Length == 0)